Remove call to LockStatus
LockStatus is returned as an object within LockDetails, however it is missing the convenience 'state' data. So we need to pull out the LockStatus and use the August-api function to add the state data.

Also, retry count isn't available, so we need another solution to determining if the lock is jammed. Though that seems minor to reducing the number of calls on the August API.